Revisio containers Nou programa de demostracio SpiOledDemo

+++ v2/Include/Services/Gui/Visuals/eosButtonBase.h
@@ -26,7 +26,7 @@ namespace eos {
 			struct EventArgs {
 				ButtonBase *button;
 				EventType event;
-				uint8_t id;
+				unsigned id;
 			};
 			typedef ICallbackP1<const EventArgs&> IEventCallback;
